requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.101.1
h11==0.16.0
idna==3.10
lxml==5.4.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
reportlab==4.4.0
requests==2.32.3
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
XlsxWriter==3.2.3
